Create Scala 2 veneer for SourceInfo
The SourceInfo hierarchy is sealed, and very basic, so most of the code
is just moved to src/main/scala-2.
